requirements.txt
alembic>=1.13.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
python-telegram-bot[job-queue]>=21.0
SQLAlchemy>=2.0.0
